Renames after morphology -> image.models move
