+++ gammapy/astro/population/simulate.py
@@ -11,7 +11,7 @@ from ...utils import coordinates as astrometry
 from ...utils.coordinates import D_SUN_TO_GALACTIC_CENTER
 from ...utils.distributions import draw, pdf
 from ...utils.random import sample_sphere, sample_sphere_distance, get_random_state
-from ...morphology.models import morph_types
+from ...image.models import morph_types
 from ..source import SNR, SNRTrueloveMcKee, PWN, Pulsar
 from ..population import Exponential, FaucherSpiral, RMIN, RMAX, ZMIN, ZMAX, radial_distributions
 from ..population import VMIN, VMAX, velocity_distributions
