+++ gammapy/image/models/tests/test_model.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import absolute_import, division, print_function, unicode_literals
 from astropy.utils.data import get_pkg_data_filename
-from ...extern import xmltodict
+from ....extern import xmltodict
 
 
 def test_model_xml_read_write():
